Create rhino.inside nuget package
Build the nuget package with `msbuild -t:"restore;pack" -p:"Configuration=Release"`

RhinoCommon, Grasshopper and RhinoWindows are dependencies of the Rhino.Inside package and will be automatically available to any project in which the package is installed (via "transitive dependencies"). However, RhinoInside.dll is the only assembly that will be copied to the build output directory.

Fixes #196
Fixes RH-54969